updated the user.py file  authentication and delete function

+++ views/users.py
@@ -9,10 +9,8 @@ user_bp = Blueprint("user_bp", __name__)
 #get all users
 @user_bp.route('/users', methods = ['GET'])
 def get_all_users():
-    #Authentication 
     # get all the users 
     users = User.query.all()
-
     #create an empty list to store the users 
     user_list= []
     
@@ -29,7 +27,6 @@ def get_all_users():
 @user_bp.route('/users/<int:user_id>', methods = ['GET'])
 def get_user_id(user_id):
     user = User.query.get(user_id)
-    # users = User.query.filter_by(user_id = current_user_id)
     if user:
         return jsonify({  
         "id":user.id,
@@ -57,11 +54,10 @@ def get_user_id(user_id):
     else: 
         return jsonify({"Error": "User does not exist"})
     
-#Create a User  
+#Create /post a User 
 @user_bp.route('/users', methods = ["POST"])
-@jwt_required()
+# No need to create an authentication since you are posting. 
 def post_user_id():
-    current_user_id=get_jwt_identity()
     # get the data
     data = request.get_json()
     name = data["name"]
@@ -85,16 +81,17 @@ def post_user_id():
         return jsonify({"Success": "User added successfully"}), 201
     
 #Update a User  
-@user_bp.route('/users/<user_id>', methods = ["PATCH","PUT"])
-def update_user_id(user_id):
-    
+@user_bp.route('/users/update', methods = ["PATCH","PUT"])
+@jwt_required()
+def update_user():
+    current_user_id = get_jwt_identity()
     # user will be none if no user is found
     # get all the Users 
-    user= User.query.get(user_id)
+    user= User.query.get(current_user_id)
     
-    # check if the user exist, 
+    # check if not the current_user
     if not user:
-        return jsonify({"Error": "Use not found based on the id. Choose an existing user-id"}), 404
+        return jsonify({"Error":"Login to Update/ User not Found"}), 404
     
  #if the data is not provided issues the data
     data = request.get_json()
@@ -102,38 +99,44 @@ def update_user_id(user_id):
     email = data.get("email", user.email)
     password = data.get("password", user.password)
     is_admin= data.get("is_admin", user.is_admin)    
-      
-    #check if the data is the same/identical no change was made 
-    if name==user.name and email==user.email and  is_admin ==user.is_admin:
-        return jsonify({"Error": "User full data is in the database, update something else"}), 400
     
-    # check for existing name or email if they already exist in another user 
-    check_name = User.query.filter(User.name==name ,User.id!=user_id).first()
-    check_email= User.query.filter(User.email==email , User.id!=user_id).first()
+    
+    # user= User.query.get(current_user_id)
+    # check if the name already exist in the database 
+    check_name = User.query.filter_by(name=name and id!=user).first()
+    check_email= User.query.filter_by(email=email and id!=user).first()
         
     #if the user with name or email exist 
     if check_name and check_email:
-        return jsonify({"Error": "A user with this name or email already exist. Update a different name or email or something else"}),406
+        return jsonify({"Error": "Name or Email already exist. Update a different name or email or something else"}),406
+    
+    #check if the data is the identical to the current_user/ and nothing has been changed.
+    if name==user.name and email==user.email and is_admin==user.is_admin:
+        return jsonify({"Error":"No change detected in your update"}), 400
+    
     else: 
         #if no conflict update data
             user.name = name 
             user.email = email
-            user.password = password
             user.is_admin = is_admin
+            if password:
+                user.password = generate_password_hash(password)
         
             #commit the function 
             db.session.commit()
-            return jsonify({"Success": f"User with  ID {user_id} was updated successfully"}),200
+            return jsonify({"Success": f"User with  ID {current_user_id} was updated successfully"}),200
         
-        ## Hash the password if a new one is provided
-    # if password:
-        # user.password = generate_password_hash(password)
-    
-#Delete User   
-@user_bp.route('/users/<int:user_id>' ,methods=['DELETE']) 
-def delete_user(user_id):
+
+#User Can Delete there own Accounts
+@user_bp.route('/users/delete' ,methods=['DELETE']) 
+@jwt_required()
+def delete_user():
     #get the all the users
-    user = User.query.get(user_id)
+    current_user_id = get_jwt_identity()
+    user = User.query.get(current_user_id)
+    
+    if not user:
+        return jsonify({"error": "Delete your own account/ User not found "})
     if user:
         db.session.delete(user)
         db.session.commit()
@@ -141,18 +144,29 @@ def delete_user(user_id):
     else:
          return jsonify({"Error": "User does not exist"}), 406
      
-# Delete All users / Can be done by the admin 
 
-@user_bp.route('/users', methods=['DELETE']) 
+# Deleting all users can be done by the admin only 
+@user_bp.route('/users/<int:user_id>/delete', methods=['DELETE']) 
 @jwt_required()      
-def delete_all_user():
+def delete_all_user(user_id):
     current_user_id = get_jwt_identity()
-    #get the all the users
-    user = User.query.delete()
-    db.session.commit()
-    return jsonify({"Success":"Users deleted successfully"}), 201
-
-
-      
-        
+    current_user = User.query.get(current_user_id) # get the admin id. 
+    # get the all the user
+    if not current_user or not current_user.is_admin:
+        return jsonify({"Error": "You are not an admin to delete an account "}), 403 
     
+    user_delete= User.query.get(user_id)
+    if not user_delete: 
+        return jsonify({"Error": "User not Found"})
+    #if the user is the admin then no deletion can happen
+    if user_delete.id == current_user.id:
+        return jsonify({"Error": "You cannot delete your own account"})
+    else: 
+        db.session.delete(user_delete)
+        db.session.commit()
+        return jsonify({"Success":f"User with {user_id}  has been deleted successfully"}), 200
+    
+
+    
+    
+
